requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
sounddevice
numpy
